+++ example/src/main/java/org/blockstack/android/MainActivity.kt
@@ -1,6 +1,7 @@
 package org.blockstack.android
 
 
+import android.annotation.SuppressLint
 import android.content.Intent
 import android.graphics.Bitmap
 import android.graphics.BitmapFactory
@@ -19,7 +20,7 @@ import org.jetbrains.anko.coroutines.experimental.bg
 import java.io.ByteArrayOutputStream
 import java.net.URL
 
-
+@SuppressLint("SetTextI18n")
 class MainActivity : AppCompatActivity() {
     private val TAG = MainActivity::class.java.simpleName
 
@@ -33,41 +34,45 @@ class MainActivity : AppCompatActivity() {
         setContentView(R.layout.activity_main)
         setSupportActionBar(toolbar)
 
-        signInButton.isEnabled = false
-        getStringFileButton.isEnabled = false
-        putStringFileButton.isEnabled = false
-
         val config = java.net.URI("https://flamboyant-darwin-d11c17.netlify.com").run {
             org.blockstack.android.sdk.BlockstackConfig(
                     this,
-                    java.net.URI("${this}/redirect"),
-                    java.net.URI("${this}/manifest.json"),
+                    "/redirect",
+                    "/manifest.json",
                     kotlin.arrayOf(org.blockstack.android.sdk.Scope.StoreWrite))
         }
 
-        _blockstackSession = BlockstackSession(this, config,
-                onLoadedCallback = {
-                    // Wait until this callback fires before using any of the
-                    // BlockstackSession API methods
 
-                    signInButton.isEnabled = true
-                })
+        _blockstackSession = BlockstackSession(this@MainActivity, config)
+        signInButton.isEnabled = true
+        getUserAppFileUrlButton.isEnabled = true
 
+        validateProofsButton.isEnabled = false
         getStringFileButton.isEnabled = false
         putStringFileButton.isEnabled = false
         getImageFileButton.isEnabled = false
         putImageFileButton.isEnabled = false
         getStringFileFromUserButton.isEnabled = false
+        getAppBucketUrlButton.isEnabled = false
+
 
         signInButton.setOnClickListener { _: View ->
-            blockstackSession().redirectUserToSignIn { userDataResult ->
-                if (userDataResult.hasValue) {
-                    Log.d(TAG, "signed in!")
+            blockstackSession().redirectUserToSignIn { errorResult ->
+                if (errorResult.hasErrors) {
+                    Toast.makeText(this, "error: " + errorResult.error, Toast.LENGTH_SHORT).show()
+                }
+            }
+        }
+
+        validateProofsButton.setOnClickListener { _ ->
+            validateProofsText.text = "Validating..."
+            val it = blockstackSession().loadUserData()
+            it?.let {
+                val ownerAddress = it.decentralizedID.split(":")[2]
+                blockstackSession().validateProofs(it.profile!!, ownerAddress, it.json.optString("username")) { result ->
                     runOnUiThread {
-                        onSignIn(userDataResult.value!!)
+                        validateProofsText.text = "${result.value?.size} proofs found."
                     }
-                } else {
-                    Toast.makeText(this, "error: " + userDataResult.error, Toast.LENGTH_SHORT).show()
                 }
             }
         }
@@ -180,9 +185,38 @@ class MainActivity : AppCompatActivity() {
             }
         }
 
-        if (intent?.action == Intent.ACTION_VIEW) {
-            handleAuthResponse(intent)
+        getAppBucketUrlButton.setOnClickListener { _ ->
+            getAppBucketUrlText.text = "Getting url ..."
+            val it = blockstackSession().loadUserData()
+            it?.let {
+                blockstackSession().getAppBucketUrl(it.hubUrl, it.appPrivateKey) { result ->
+                    runOnUiThread {
+                        getAppBucketUrlText.text =
+                                if (result.hasValue) {
+                                    result.value
+                                } else {
+                                    result.error
+                                }
+                    }
+                }
+            }
         }
+
+        getUserAppFileUrlButton.setOnClickListener { _ ->
+            getUserAppFileUrlText.text = "Getting url ..."
+            val username = "dev_android_sdk.id.blockstack";
+            val zoneFileLookupUrl = "https://core.blockstack.org/v1/names"
+            blockstackSession().getUserAppFileUrl(textFileName, username, "https://flamboyant-darwin-d11c17.netlify.com", zoneFileLookupUrl) {
+                runOnUiThread {
+                    getUserAppFileUrlText.text = if (it.hasValue) {
+                        it.value
+                    } else {
+                        it.error
+                    }
+                }
+            }
+        }
+
     }
 
     private fun onSignIn(userData: UserData) {
@@ -190,11 +224,13 @@ class MainActivity : AppCompatActivity() {
         showUserAvatar(userData.profile?.avatarImage)
         signInButton.isEnabled = false
 
+        validateProofsButton.isEnabled = true
         getStringFileButton.isEnabled = true
         putStringFileButton.isEnabled = true
         putImageFileButton.isEnabled = true
         getImageFileButton.isEnabled = true
         getStringFileFromUserButton.isEnabled = true
+        getAppBucketUrlButton.isEnabled = true
     }
 
     private fun showUserAvatar(avatarImage: String?) {
@@ -222,14 +258,13 @@ class MainActivity : AppCompatActivity() {
         Log.d(TAG, "onNewIntent")
 
         if (intent?.action == Intent.ACTION_MAIN) {
-            blockstackSession().loadUserData { userData ->
-                if (userData != null) {
-                    runOnUiThread {
-                        onSignIn(userData)
-                    }
-                } else {
-                    Toast.makeText(this, "no user data", Toast.LENGTH_SHORT).show()
+            val userData = blockstackSession().loadUserData()
+            if (userData != null) {
+                runOnUiThread {
+                    onSignIn(userData)
                 }
+            } else {
+                Toast.makeText(this, "no user data", Toast.LENGTH_SHORT).show()
             }
         } else if (intent?.action == Intent.ACTION_VIEW) {
             handleAuthResponse(intent)
@@ -245,7 +280,8 @@ class MainActivity : AppCompatActivity() {
             if (authResponseTokens.size > 1) {
                 val authResponse = authResponseTokens[1]
                 Log.d(TAG, "authResponse: ${authResponse}")
-                blockstackSession().handlePendingSignIn(authResponse, { userDataResult ->
+                blockstackSession().handlePendingSignIn(authResponse) { userDataResult: Result<UserData> ->
+
                     if (userDataResult.hasValue) {
                         val userData = userDataResult.value!!
                         Log.d(TAG, "signed in!")
@@ -255,7 +291,7 @@ class MainActivity : AppCompatActivity() {
                     } else {
                         Toast.makeText(this, "error: " + userDataResult.error, Toast.LENGTH_SHORT).show()
                     }
-                })
+                }
             }
         }
     }